Fix pxPortInitialiseStack to match a normal context save
portCONTEXT_SIZE was bumped in our CHERI port, but pxPortInitialiseStack
was calculating the old value in various stages. Instead, rework it to
look like a normal context save, ensuring it allocates the right amount
of space in the process. This should fix overflowing the top of stack
allocations as soon as they are first used, which tramples over adjacent
heap metadata.