Move trap handler installation to boot.S.
This is necessary because some CWE tests run in main instead of in a
task and were causing traps before the handler was installed.  Note
that this means that the handler has to check that pxCurrentTCB is not
NULL before proceeding as it will not be set until the scheduler is
called for the first time.