Avoid wedging bluespec_p3 in PLIC_init
Other cores suppress AXI write error responses, but bluespec_p3 does not
forward an MMIODataPRs in that instance and so the pipeline stalls
forever waiting for the response and must be reset (but the debug module
cannot do so due to the core wedging). This should do one of cause an
NMI, cause a store access fault or be ignored, as wedging the core is
not useful behaviour.

In the case of FreeRTOS, it tries to write to the 0th PLIC priority
register, which is reserved due to there being no source 0 and gets a
write error response from the Bluespec PLIC implementation. We should
therefore just start from 1 rather than 0, as regardless of whether it
causes the core to wedge we should still not be attempting to write to
reserved registers. We also have an off-by-one error as num_sources
includes source 0.

The source enabled zeroing does not suffer from including source 0
because each entry is 1 bit rather than a whole byte, thus the 0th entry
is within a non-reserved register, and the SiFive spec makes it RW but
hard-wired to 0. However, we do have an off-by-one error in the
rounding-up code.

We also seem to be able to wedge bluespec_p3 by performing unaligned
single-byte accesses to the PLIC. Given that the PLIC assumes any access
is aligned and 32-bit (it will both ignore the LSBs of the address and
use 32 wdata bits), strange things could happen anyway on other Bluespec
cores when not doing so. Since all the interrupt sources we generally
care about are in the first byte this probably doesn't matter on other
cores at the moment, but even if this didn't wedge bluespec_p3 we
shouldn't be relying on it.