FreeRTOS-libdl: Implement a new way of resolving external relocations
The old way of resolving external relocations is assuming all object globals
are added to a single publicly accessible (by every object) list, then resolve
the relocation against that. This has now
changed by restricting each object globals to itself and adding extra interface
list that holds the allowed symbols for use by other compartment.

If an object has unresolved external relocations, it means it's likely trying
to do an inter-compartment calls/references. The planned compartmentalized way
to do this is to see if that object has a capability (or is allowed) to do
such call. This is yet to be implemented and it depends on a specified policy
by the one describing the system compartments and their interactions.

If that check succeeds, copy the symbol/capability from the object that owns it to
the target object that references it, then do a relocation against the copied
one.