FreeRTOS-libdl: Mint capabilities to global FreeRTOS syms into each object
On-demend mininting for global FreeRTOS/libc symbols that an object accesses.
This makes it easier to maintain a single captable for each object refered to by
a single register (currently $cgp) and caps accessed relative to it.

No other accesses are allowed if the captable doesn't have a capability to
a symbol.