FreeRTOS-libdl/cheri: Replace AUIPCCs to load a data cap from
Workaround to do $cgp relative loads part of ccall prologues. This is until we
have a compiler support for $cgp loads. Currently there are no more AUIPCC
instructions and all caps are $cgp relative. Re-use AUIPCC instruction slot and
re-write it in order to do a $cgp-relaive load of the data cap to be passed to
a ccall in $tp. There's a limitation that the offset be within 4K at the moment,
extra instructions will be needed if we need to reference more capabilities that
do not fit in a 4K captable.