cheri-riscv/compartments: Add a separate portASM.S just for CHERI and compartments
This commit adds a separate portASM.S just for CHERI with compartmentalization
support. The diff is:
* Formatting, white-space clean-ups and make the code more consistent and readable.
* Only CHERI-support here, vanilla RISC-V should use the original portASM.S.
* s/sandboxing/compartments with design/implementation changes described later.
* Disable sandboxing trap handling/assumptions as they conflict with compartmentalization
support. Demos that need sandboxing can use the original portASM.S.

(WIP) Design/Implementation changes for compartmentalization support:
* The register-file context is saved in the caller's stack before switching to
the callee's stack.
* CCall ISR handling executes entirely in the caller's stack before switching to
the callee compartment.
* mscratch is used to hold the currently running compartment ID. CompID of value
4095 means the current thread runs uncompartmentalized of its FreeRTOS/libc.
* FreeRTOS return-data and function after a compartment is called is sealed with
a CompID of 4095. Other compartment IDs start from 0 up to 4095 (just to fit in
an single immediate load, though there is no reason not to discard that restriction).
* The stack of the callee's compartment is assumed to be saved in the first slot
of its cap table (which passed to the xPortCompartmentEnter as pxData). This also
will change once we support compartment reentrancy.
* The stack of the compartment's caller is sealed in saved in the callee-saved cs1
register to easily (and hopefully) get it back once the compartment returns.
* processed_source restores context from the stack in the case of (nesting)
compartments.
* FIXME pxPortCompartmentReturnData is used to hold the caller's stack and iif
the compartment traps, the context is restored from pxPortCompartmentReturnData.
That only works with a single leaf compartment trap right now and will need to
be fixed/changed.