FreeRTOS-libvirtio/cheri: Derive avail and used ring caps from the allocated-queue mem
The whole queue is allocated and assigned to vq->desc (being the start of the queue)
and then pointer arithmetic for used/avail is done to calc their start addresses.
This kind of pointer arithmetic results in untagged caps.

This commit derives bounded caps for avail and used rings part of the queue.