FreeRTOS-libvirtio/cheri: Derive a cap for received RX buffers from device
The buffers RX queue uses are allocated during queue init as a big chunk of
contiguous memory, and hence a cap.

The virtio device isn't capability-aware, and hence buffer addresses are 64-bit
integers. The commit gets that integer address and derives a cap for it from the
the previously allocated buf_mem.