RISC-V-Generic: Change the default DNS server to 8.8.8.8
The default QEMU's user networking uses 10.0.2.3 for a DNS server. This returns
DNS reply packet with TTL=0, and hence the DNS cache will always need to be
refreshed.

Using 8.8.8.8 as a remote DNS server makes packets go through the host/gateway and
gets a proper TTL time for DNS replies.