FreeRTOS-libvirtio: virtio-net - TX - Dedicate a single always-zero virtio header
The virito header used in this library is always zero and doesn't change currently.
Instead of having to memset it to zero on every xmit, just allocate a single one
that gets zeroed in the BSS and use it for every packet xmit.

This will need to be changed if/when the driver/device support more features.